Reduce the count size
Count can be at most one billion, so 32 bits suffice.

This makes the Entry struct 160 bytes, which is still 2.5 cache lines,
so not quite small enough to fit into two.

+++ src/main.rs
@@ -148,5 +148,5 @@ struct Entry {
     min: f32,
     max: f32,
     sum: f64,
-    count: usize,
+    count: u32,
 }
